+++ ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/retry/retry.h
@@ -31,6 +31,7 @@ struct TRetryOperationSettings {
     }
     FLUENT_SETTING_FLAG(Verbose);
     FLUENT_SETTING_FLAG(RetryUndefined);
+    FLUENT_SETTING_FLAG(CatchYdbExceptions);
 
     static TBackoffSettings DefaultFastBackoffSettings() {
         return TBackoffSettings()